requirements.txt
llama-index
python-dotenv==1.0.1
pandas
openpyxl
openai